backend/requirements/common.txt
django~=4.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Pillow==9.2.0
django-cors-headers==3.13.0
django-mptt==0.13.4